+++ lib/flop/custom_types/like.ex
@@ -24,5 +24,5 @@ defmodule Flop.CustomTypes.Like do
   def type, do: :string
   def load(_), do: :error
   def dump(_), do: :error
-  # coveralls-ignore-end
+  # coveralls-ignore-stop
 end