+++ test/flop/filter_test.exs
@@ -16,7 +16,7 @@ defmodule Flop.FilterTest do
   end
 
   describe "allowed_operators/1" do
-    test "returns a list of operators for each native Ecto type" do
+    test "returns a list of operators for each Ecto type" do
       types = [
         :id,
         :binary_id,
@@ -36,12 +36,54 @@ defmodule Flop.FilterTest do
         :naive_datetime_usec,
         :utc_datetime,
         :utc_datetime_usec,
-        {:parameterized, Ecto.Enum, type: :string}
+        {:parameterized, {Ecto.Enum, %{type: :string}}},
+        {:ecto_enum, [:one, :two]},
+        {:from_schema, MyApp.Pet, :mood}
       ]
 
       for type <- types do
-        assert [op | _] = Filter.allowed_operators(type)
+        assert [op | _] = ops_for_type = Filter.allowed_operators(type)
         assert is_atom(op)
+
+        ops_for_field =
+          Filter.allowed_operators(%Flop.FieldInfo{ecto_type: type})
+
+        assert ops_for_type == ops_for_field
+      end
+    end
+
+    test "returns list of operators for enum" do
+      types = [
+        # by internal representation Ecto < 3.12.0
+        {:parameterized, Ecto.Enum, %{type: :string}},
+        # by internal representation Ecto >= 3.12.0
+        {:parameterized, {Ecto.Enum, %{type: :string}}},
+        # same with init function
+        Ecto.ParameterizedType.init(Ecto.Enum, values: [:one, :two]),
+        # by convenience format
+        {:ecto_enum, [:one, :two]},
+        # by reference
+        {:from_schema, MyApp.Pet, :mood}
+      ]
+
+      expected_ops = [
+        :==,
+        :!=,
+        :empty,
+        :not_empty,
+        :<=,
+        :<,
+        :>=,
+        :>,
+        :in,
+        :not_in
+      ]
+
+      for type <- types do
+        assert Filter.allowed_operators(type) == expected_ops
+
+        assert Filter.allowed_operators(%Flop.FieldInfo{ecto_type: type}) ==
+                 expected_ops
       end
     end